+++ src/yt_config.h
@@ -62,9 +62,9 @@ bool yt_config_redraw_repairs(struct yt_database *database,
 	const struct yt_record *field, float working_maximum_holds,
 	struct yt_record *result, struct yt_error *error);
 
-int yt_date_serial(const struct yt_clock_value *date, float epoch_year,
+int yt_date_serial(const struct yt_clock_value *date, uint8_t epoch_year,
     int *adjusted_year);
-bool yt_current_date_serial(const struct yt_clock *clock, float epoch,
+bool yt_current_date_serial(const struct yt_clock *clock, uint8_t epoch,
     int *serial, int *adjusted_year, struct yt_error *error);
 void yt_format_date(const struct yt_clock_value *value, char dest[11]);
 void yt_format_time(const struct yt_clock_value *value, char dest[9]);
